refactor: move TableRenderer.tsx from utils to components directory
    - Relocated TableRenderer.tsx to components for better organization within the project structure
